Testnet 2.2 (#54)
* Removing DB LOCK

* Pending state

* Clean up

* Added http endpoint to node, changed calling functions

* Proto in line

* Clean up

* DG-705, DG-691, DG-707, DG-708, DG-711

* Clean up

* Clean up

* DG-626 adding rumor timeouts

* DG-682

* DG-719

* Clean up

* Clean up

* adding tx receive timeout

* DG-720 fixed exists map concurrency for queues

Dumb mistake.  Did not have a thread safe map implementation so added one with a concurrency unit test

* Clean up

* Added authenticate struct

* Update

* Clean up

* adding timeout check just before executing the tx

* adding abi helper

* added more to abi helper

Still need more code to properly handle Arrays

* initial

switching computers, checking in

* moved timeouts to constants removing hard-coded values

We are currently doing all of the calculations in milliseconds.  I think we need to change this to nano-seconds for back-end code. This will be particularly helpful with the delegate to delegate comms.

* incremental checkin for abi_helper

added some extra logging to help figuring out this array thing.

* set default isBookkeeper to true

* default isBookkeeper to true

* increase timeout to 2 seconds

* adding some logging for testing

* upping timeout values to unblock QA on functional testing

While I do research, want to just eliminate this as an issue.

* updated to latest protobuf to see if it fixes the issue

* adding some logging for QA to do some checking

* more informative logging for QA

* Dg-690 (#55)

* New API functionality and bug fixes

* backwards compatible, negative not supportive on page

* 690: Auto stash before merge of "DG-690" and "testnet-2.2"

* hide

* added back receipts endpoint

also logging grpc key during sync process.  A key is being created somewhere that is invalid.

* fixed abi helper so it now supports array parameters

* DG-762

fixed grpc errors when disgo starts.  Caused by pulling DVM cache values which should not be part of the sync process.

* added logging to file and pretty print for config

* trying with 100% consensus

* Back to 2/3 ... issue was identified