modified Gossip to not gossip so poorly (#124)
The gossipworker was not efficiently handling gossip and we had an overload of gossiping that was really problematic.  Big change was to separate who delegates have sent to and who they received from. Those are handled enirely separate now so that we can validate on recieve and not send to the same delegate many times.