Added explicit processing to the DVM result for []byte return values
since it is automatically converted to base64, it must be decoded and I found that the VM adds null termination, so that must be trimmed from the end as well.  One important note is to make sure that "string" is the expected result.  I have this feeling that this issue might have cases where we want to keep the base64 or the base64 bytes.