Second instance of this branch
The number of merges into this  branch made it extremely difficult to isolate the code changes.  I have have reproduced the branch as a single check in to create a PR from.
I do think we need to talk this over since it does have implications to how the system behaves.  I think delegates will need to behave differently, but we can talk that over.