updated the default values to reasonable numbers (#180)
The min hertz/gas was set to 1 trillion, so an intdeterminate loop could run seemingly forever.  I lowered to a reasonable number after some testing.
High priority for post-main-net is to add hert limits to the initial transaction request and at least set the limit to available hertz if nothing is specified.